requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.1
Automat==22.10.0
billiard==3.6.4.0
boto3==1.26.59
botocore==1.29.59
celery==5.2.7
cffi==1.15.1
channels==4.0.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==39.0.0
daphne==4.0.0
dateparser==1.1.6
Django==4.1.3
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-filter==22.1
django-rest-passwordreset==1.3.0
django-storages==1.13.2
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
h11==0.14.0
humanize==4.4.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jmespath==1.0.1
kombu==5.2.4
Markdown==3.4.1
maya==0.6.1
pendulum==2.1.2
Pillow==9.4.0
prompt-toolkit==3.0.36
psycopg2==2.9.5
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.0.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.7
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
redis==4.4.2
regex==2022.10.31
s3transfer==0.6.0
service-identity==21.1.0
six==1.16.0
snaptime==0.2.4
sqlparse==0.4.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
uvicorn==0.20.0
vine==5.0.0
wcwidth==0.2.6
zope.interface==5.5.2
